+++ src/main/resources/ChallengeSampleDataSet1.txt
